Extract an OpmlImportator service from the job
This makes the OpmlImportator easier to use.
