+++ src/services/OpmlImportatorError.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace flusio\services;
+
+/**
+ * Exception raised by the OpmlImportator service.
+ *
+ * @see \flusio\services\OpmlImportator
+ *
+ * @author Marien Fressinaud <dev@marienfressinaud.fr>
+ * @license http://www.gnu.org/licenses/agpl-3.0.en.html AGPL
+ */
+class OpmlImportatorError extends \RuntimeException
+{
+}
