+++ src/services/UserCreator.php
@@ -60,25 +60,14 @@ class UserCreator
         // Load default feeds
         $default_opml_filepath = \Minz\Configuration::$data_path . '/default-feeds.opml.xml';
         if (file_exists($default_opml_filepath)) {
-            $importations_filepath = \Minz\Configuration::$data_path . '/importations';
-            if (!file_exists($importations_filepath)) {
-                @mkdir($importations_filepath);
+            try {
+                $opml_importator_service = new OpmlImportator($default_opml_filepath);
+                $opml_importator_service->importForUser($user);
+            } catch (OpmlImportatorError $e) {
+                \Minz\Log::error("Error while importing default feeds for user {$user->id}: {$e->getMessage()}");
+                // Don't pass the error to the parent as it's a "minor" issue
+                // (the user actually exists and is functional)
             }
-
-            $user_opml_filepath = "{$importations_filepath}/opml_{$user->id}.xml";
-            $copy_is_ok = copy($default_opml_filepath, $user_opml_filepath);
-            if (!$copy_is_ok) {
-                \Minz\Log::error("Default OPML file can’t be copied to {$user_opml_filepath}");
-                return $user;
-            }
-
-            $importation = models\Importation::init('opml', $user->id, [
-                'opml_filepath' => $user_opml_filepath,
-            ]);
-            $importation->save();
-            $importator_job = new jobs\OpmlImportator();
-            $importator_job->perform($importation->id);
-            models\Importation::delete($importation->id);
         }
 
         return $user;
